+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 - Added a report builder to the dashboard to view multiple artifacts at
   once ([#210](https://gitlab.heigit.org/climate-action/web-app/-/issues/210))
 - Users can now request a demo computation, if the plugin provides that functionality
+- Computations can now be shared and imported between users ([#116](https://gitlab.heigit.org/climate-action/web-app/-/issues/116))
 
 ## [2.1.0](https://gitlab.heigit.org/climate-action/web-app/-/releases/2.1.0) - 2025-01-29